+++ src/features/connector/connectorApiSlice.ts
@@ -83,6 +83,26 @@ export interface DecentralIdentityUrlsType {
   decentralIdentityManagementAuthUrl: string
 }
 
+export interface ConnectorDetailsType {
+  name: string
+  location: string
+  id: string
+  type: string
+  status: string
+  hostId: string
+  hostCompanyName: string
+  connectorUrl: string
+  technicalUser: string
+  selfDescriptionDocumentId?: string | null
+}
+
+export type updateConnectorUrlType = {
+  connectorId: string
+  body: {
+    connectorUrl: string
+  }
+}
+
 export const apiSlice = createApi({
   reducerPath: 'rtk/admin/connector',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
@@ -100,6 +120,9 @@ export const apiSlice = createApi({
         method: 'POST',
         body,
       }),
+    }),
+    fetchConnectorDetails: builder.query<ConnectorDetailsType, string>({
+      query: (connectorID) => `/api/administration/Connectors/${connectorID}`,
     }),
     deleteConnector: builder.mutation<void, string>({
       query: (connectorID) => ({
@@ -138,15 +161,24 @@ export const apiSlice = createApi({
     fetchDecentralIdentityUrls: builder.query<DecentralIdentityUrlsType, void>({
       query: () => 'api/administration/companydata/decentralidentity/urls',
     }),
+    updateConnectorUrl: builder.mutation<void, updateConnectorUrlType>({
+      query: (data) => ({
+        url: `/api/administration/Connectors/${data.connectorId}/connectorUrl`,
+        method: 'PUT',
+        body: data.body,
+      }),
+    }),
   }),
 })
 
 export const {
   useCreateConnectorMutation,
   useCreateManagedConnectorMutation,
+  useFetchConnectorDetailsQuery,
   useDeleteConnectorMutation,
   useFetchConnectorsQuery,
   useFetchManagedConnectorsQuery,
   useFetchOfferSubscriptionsQuery,
   useFetchDecentralIdentityUrlsQuery,
+  useUpdateConnectorUrlMutation,
 } = apiSlice