+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Technical User Creation - API Response Changes
 - AppSubscription, Service Subscription, Company Subscription
   - Validation of Technical User UIs for multiple tech user support
+- Application Request
+  - Implement Expandable Status Detail Overlay
+- Admin Credential
+  - Implement Credential Revocation Feature
+  - Add Credential Status
+- Connector Management
+  - Updated main page UI and added details overlay and functionality
 
 ### Change